+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 > This project follows the [Keep a Changelog](https://keepachangelog.com) format.
 
 ## [Unreleased]
-* add `$contains` and `$regex` value matchers for request url, query, headers and body.
-* support `RegExp` directly in value matchers as a shorthand for `{ $regex }`.
+* add `$$contains` and `$$regex` value matchers for request url, query, headers and body.
+* support `RegExp` directly in value matchers as a shorthand for `{ $$regex }`.
 
 ## [0.4.0] - 2026-06-02
 * refactor and export response builder.